Documentacion, arreglos pruebas y fase

+++ src/app/pruebas.java
@@ -77,6 +77,7 @@ class pruebas {
 		assertEquals(Math.pow(2, 0.5),num.getPartR());
 		assertEquals(Math.atan(1),num.getPartI());
 	}
+	
 	@Test
 	public void deberiaCambiarLasCoordenadasACardinales() {
 		double[] num1 = {Math.pow(2, 0.5),Math.atan(1)};
@@ -84,16 +85,18 @@ class pruebas {
 		ComplexNumber num = new ComplexNumber(num1[0],num1[1]);
 		num.cambiarForma('P');
 		Library lb = new Library(num2,num2) ;
-		lb.cambiarForma(num);	
-		assertEquals(1,num.getPartR());
-		assertEquals(1,num.getPartI());
+		lb.cambiarForma(num);
+		assertEquals(1,(int)(num.getPartR()));
+		assertEquals(1,(int)(num.getPartI()));
 	}
+	
 	@Test
 	public void deberiaDarFaseDeUnNumeroCartesianas() {
 		int[] num1 = {1,1};
 		ComplexNumber num = new ComplexNumber(num1[0],num1[1]);
 		assertEquals(Math.atan(1),num.fase());
 	}
+	     
 	@Test
 	public void deberiaDarFaseDeUnNumeroPolares() {
 		int[] num1 = {60,1};
